fix: standardize extent_id to u32 on wire and sync design doc with code
Issue #10 — extent_id u32/u64 width mismatch:
- Change client seal()/extent_node_seal() return type from (u64, String)
  to (u32, String) since ExtentId is u32
- Remove redundant `as u32` casts at all call sites (examples, tests, benches)
- Fix test in payload.rs that wrote extent_id as u64 instead of u32
- Fix stale comments referencing extent_id as u64 in store.rs, extent.rs,
  and design.md

Issue #12 — heartbeat/lifecycle wire format doc mismatch:
- Update CONNECT (0x10): document request_id in variable header, node_id +
  addr + interval in payload section (matching code)
- Update CONNECT_ACK (0x11), DISCONNECT_ACK (0x13): document request_id in
  variable header (matching code)
- Update DISCONNECT (0x12): document request_id in variable header, node_id
  in payload section (matching code)
- Update HEARTBEAT (0x14): document request_id in variable header, metrics
  in payload; document response format with request_id (matching code)

Issue #7 — describe_stream ordering ambiguity:
- Document DESCRIBE_STREAM_RESP returns extents ordered by extent_id
  descending (latest first)

+++ benches/concurrent_multi_stream.rs
@@ -101,7 +101,7 @@ async fn client_task(
                     .await
                     .unwrap_or_else(|e| panic!("client {client_id}: seal failed: {e}"));
 
-                extent_id = ExtentId(new_extent_id_raw as u32);
+                extent_id = ExtentId(new_extent_id_raw);
                 primary_addr = new_primary_addr;
                 seal_count += 1;
